Update login related components
- move them to login folder
- start implementing new look for login
- add some svgs (duck + google login)
