--- frontend/src/components/GoogleLogin.css
@@ -1,30 +1,0 @@
-.GoogleLogin {
-    text-align: left;
-}
-
-.GoogleLogin-logo {
-    height: 18dp;
-    padding-left: 18dp;
-    padding-top: 8dp;
-    padding-bottom: 8dp;
-}
-
-.GoogleLogin-logo:hover {
-    background-color: #707070;
-}
-
-.GoogleLabel {
-    font-size: 14px;
-    justify-content: center;
-}
-
-.GoogleButton {
-    height: 40dp;
-    border-color: #e3e3e3;
-    padding: 0;
-    border: none;
-    background: none;
-    align-items: center;
-    justify-content: center;
-    position: relative;
-}
